Move `ComponentProvider` and related classes to API
