+++ src/Config/SDK/Configuration/ComponentProvider.php
@@ -2,39 +2,8 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Config\SDK\Configuration;
+namespace OpenTelemetry\API\Configuration\Config;
 
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Builder\NodeBuilder;
+use function interface_exists;
 
-/**
- * A component provider is responsible for interpreting configuration and returning an
- * implementation of a particular type.
- *
- * @template T
- */
-interface ComponentProvider
-{
-    /**
-     * @param array $properties properties provided for this component provider
-     * @param Context $context context that should be used to resolve component plugins
-     * @return T created component, typehint has to specify the component type that is
-     *         provided by this component provider
-     *
-     * @see ComponentPlugin::create()
-     */
-    public function createPlugin(array $properties, Context $context): mixed;
-
-    /**
-     * Returns an array node describing the properties of this component provider.
-     *
-     * @param ComponentProviderRegistry $registry registry containing all available component providers
-     * @param NodeBuilder $builder node builder used to create configuration nodes
-     * @return ArrayNodeDefinition array node describing the properties
-     *
-     * @see ComponentProviderRegistry::component()
-     * @see ComponentProviderRegistry::componentList()
-     * @see ComponentProviderRegistry::componentNames()
-     */
-    public function getConfig(ComponentProviderRegistry $registry, NodeBuilder $builder): ArrayNodeDefinition;
-}
+interface_exists(\OpenTelemetry\API\Configuration\Config\ComponentProvider::class);
