+++ src/SDK/composer.json
@@ -20,7 +20,7 @@
         "php": "^8.1",
         "ext-json": "*",
         "nyholm/psr7-server": "^1.1",
-        "open-telemetry/api": "~1.0 || ~1.1",
+        "open-telemetry/api": "~1.4.0",
         "open-telemetry/context": "^1.0",
         "open-telemetry/sem-conv": "^1.0",
         "php-http/discovery": "^1.14",
